Refactor classes
